refactor: migrate header config imports to CSL API
Agent-Logs-Url: https://github.com/camunda/camunda/sessions/de9065eb-1181-4f70-9b0e-ee34c7a37a3f
