+++ authentication/src/test/java/io/camunda/authentication/config/AbstractWebSecurityConfigTest.java
@@ -24,8 +24,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.authentication.config.controllers.TestApiController;
 import io.camunda.security.api.context.CamundaAuthenticationProvider;
-import io.camunda.security.configuration.headers.ContentSecurityPolicyConfig;
-import io.camunda.security.configuration.headers.PermissionsPolicyConfig;
+import io.camunda.security.api.model.config.headers.ContentSecurityPolicyConfig;
+import io.camunda.security.api.model.config.headers.PermissionsPolicyConfig;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureMockMvc;
 import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureWebMvc;
