--- security/security-core/src/main/java/io/camunda/security/configuration/headers/values/CrossOriginOpenerPolicy.java
@@ -1,14 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-package io.camunda.security.configuration.headers.values;
-
-public enum CrossOriginOpenerPolicy {
-  UNSAFE_NONE,
-  SAME_ORIGIN_ALLOW_POPUPS,
-  SAME_ORIGIN
-}
